max/requirements.txt
transformers
scikit-learn
torch
matplotlib
seaborn
numpy
pandas
datasets
azure-ai-textanalytics
lightning
setfit
transformers[torch]
sentencepiece
ollama
#transformers==4.40.2
